feat: update architeture
